chore: drops support for node 14 and commonjs BREAKING
