+++ docs/index.html
@@ -38,7 +38,7 @@
   <meta name="google-site-verification" content="K0j5zueKp7lYKMR7WihmcTJbLpHxBOpaXZEI_s0nHTw" />
   <link rel="canonical" href="https://state-machine-cat.js.org">
   <script nonce="known-inline-script">var LOG = false;</script>
-  <script src="smcat-online-interpreter.min.js" defer integrity="sha512-e2oKPc9ZKd/YYy3cZ6TN8gjz8kItJkatPVVx7RwtiAGOvOuUwth7bjo7Di+1XX+sc/eFy/Ol/qMxeXVvedlVkg&#x3D;&#x3D;"></script>
+  <script src="smcat-online-interpreter.min.js" type="module" defer integrity="sha512-343znQszNgetazsC9gzGFVNbzP/0AlqD/7zT5PpqpDjd6de/CX97JhpTnb4LJOCTgyGq5OozDcZ9nho7Kv76AA&#x3D;&#x3D;"></script>
   <script defer src="https://code.getmdl.io/1.3.0/material.min.js" async></script>
   <script async src="https://www.googletagmanager.com/gtag/js?id=G-DCPVBY7NNV"></script>
   <script nonce="known-inline-script">
