+++ docs/inpage.html
@@ -22,8 +22,9 @@
     />
     <script
       src="state-machine-cat-inpage.min.js"
+      type="module"
       defer
-      integrity="sha512-JEQsaHIIY1EEDokPVAYgIe6OUQXkupmOTV14+YwGZI92R+naocrKwOV2tdFlNLvnyb2N36D7y0fhpGNagDm5hQ&#x3D;&#x3D;"
+      integrity="sha512-IKR7WSCR1VdAUE1dIYPGjwH+TuojJ3any5jus1XEwbt2ACnOfERGcJp7Bz572/oJGrPKx91v/ZXr4cJWsqrTmQ&#x3D;&#x3D;"
     ></script>
     <style>
       body { font-family: sans-serif; margin: 0 auto; max-width: 799px;
