+++ src/render/index-node.mts
@@ -5,7 +5,7 @@ import type {
   RenderFunctionType,
   StringRenderFunctionType,
 } from "types/state-machine-cat.js";
-import smcatRendererAsImported from "./smcat/index.js";
+import smcatRendererAsImported from "./smcat/index.mjs";
 import renderDot from "./dot/index.mjs";
 import vector from "./vector/vector-native-dot-with-fallback.mjs";
 import oldVector from "./vector/vector-with-viz-js.mjs";
