+++ src/render/index.mts
@@ -5,7 +5,7 @@ import type {
   RenderFunctionType,
   StringRenderFunctionType,
 } from "types/state-machine-cat.js";
-import smcatRendererAsImported from "./smcat/index.js";
+import smcatRendererAsImported from "./smcat/index.mjs";
 import renderDot from "./dot/index.mjs";
 import svg from "./vector/vector-with-viz-js.mjs";
 import scjson from "./scjson/index.mjs";
