+++ src/render/scxml/index.mts
@@ -1,6 +1,6 @@
 import type { StringRenderFunctionType } from "types/state-machine-cat.js";
 import ast2scjson from "../scjson/index.mjs";
-import renderFomSCJSON from "./render-from-scjson.js";
+import renderFomSCJSON from "./render-from-scjson.mjs";
 
 const renderSCXML: StringRenderFunctionType = (pStateMachine) =>
   renderFomSCJSON(ast2scjson(pStateMachine));
